+++ openadapt/strategies/__init__.py
@@ -5,6 +5,7 @@ Module: __init__.py
 # flake8: noqa
 
 from openadapt.strategies.base import BaseReplayStrategy
+from openadapt.strategies.visual_browser import VisualBrowserReplayStrategy
 
 # disabled because importing is expensive
 # from openadapt.strategies.demo import DemoReplayStrategy
@@ -13,7 +14,5 @@ from openadapt.strategies.segment import SegmentReplayStrategy
 from openadapt.strategies.stateful import StatefulReplayStrategy
 from openadapt.strategies.vanilla import VanillaReplayStrategy
 from openadapt.strategies.visual import VisualReplayStrategy
-from openadapt.strategies.visual_browser import VisualBrowserReplayStrategy
-
 
 # add more strategies here
